+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
-## 1.0.0 
+## 1.0.1
+
+- Updated the logo
+- Improved code base and corrected typos
+- Updated `README.md`
+
+## 1.0.0
 
 - Initial release
-- Added two palettes: Material and Classes
+- Added two palettes: Material and Classic
 - Customizable contrast, cursor and selection colors